Refactor config creation to support both AWS and GCP
