+++ src/nimbo/core/config/gcp_config.py
@@ -1,0 +1,5 @@
+from nimbo.core.config.common import BaseConfig
+
+
+class GcpConfig(BaseConfig):
+    ...
