+++ src/nimbo/tests/config.py
@@ -1,14 +1,22 @@
+import abc
 import os
 import random
 import string
-from typing import Any, Dict
+from typing import Any, Dict, Union
 
 import pydantic
 
-from nimbo.core.config import NimboConfig, RequiredCase, load_yaml_from_file
+from nimbo.core.config import (
+    AwsConfig,
+    CloudProvider,
+    GcpConfig,
+    RequiredCase,
+    load_yaml_from_file,
+)
 
-CONDA_ENV = "env.yml"
+# TODO: support multiple config files
 NIMBO_CONFIG_FILE = "nimbo-config.yml"
+CONDA_ENV = "env.yml"
 ASSETS_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), "assets")
 
 
@@ -20,7 +28,7 @@ S3_DATASETS_PATH = f"s3://nimbo-test-bucket/{_random_string_of_len(20)}"
 S3_RESULTS_PATH = f"s3://nimbo-test-bucket/{_random_string_of_len(20)}"
 
 
-class NimboTestConfig(NimboConfig):
+class BaseTestConfig(pydantic.BaseModel, abc.ABC):
     _initial_state: Dict[str, Any] = pydantic.PrivateAttr(default_factory=dict)
     _nimbo_config_file_exists: bool = pydantic.PrivateAttr(default=True)
 
@@ -40,6 +48,12 @@ class NimboTestConfig(NimboConfig):
         for key, value in self._initial_state.items():
             setattr(self, key, value)
 
+    @abc.abstractmethod
+    def inject_required_config(self, *cases: RequiredCase) -> None:
+        ...
+
+
+class AwsTestConfig(BaseTestConfig, AwsConfig):
     def inject_required_config(self, *cases: RequiredCase) -> None:
         cases = RequiredCase.decompose(*cases)
 
@@ -68,10 +82,20 @@ class NimboTestConfig(NimboConfig):
                 return file
 
 
-def make_config() -> NimboTestConfig:
-    config = NimboTestConfig(
-        **load_yaml_from_file(os.path.join(ASSETS_PATH, NIMBO_CONFIG_FILE))
-    )
+class GcpTestConfig(BaseTestConfig, GcpConfig):
+    def inject_required_config(self, *cases: RequiredCase) -> None:
+        pass
+
+
+def make_config() -> Union[AwsTestConfig, GcpTestConfig]:
+    raw_config = load_yaml_from_file(os.path.join(ASSETS_PATH, NIMBO_CONFIG_FILE))
+    provider = CloudProvider(raw_config["provider"].lower())
+
+    if provider == CloudProvider.AWS:
+        config = AwsTestConfig(**raw_config)
+    else:
+        config = GcpTestConfig(**raw_config)
+
     config.save_initial_state()
 
     return config
